requirements.txt
Django==2.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
psycopg2-binary==2.7.4
PyJWT==1.6.4
pytz==2018.4
